refactor: insert footer links folders into folder of footer
